Add dropout scheduling in training all.

+++ model/models.py
@@ -157,7 +157,7 @@ class MelTransformer(Transformer):
                 tf.TensorSpec(shape=(None, None, decoder_postnet.mel_channels), dtype=tf.float64),
                 tf.TensorSpec(shape=(None, None, decoder_postnet.mel_channels), dtype=tf.float64),
                 tf.TensorSpec(shape=(None, None), dtype=tf.int64),
-                tf.TensorSpec(shape=(None,), dtype=tf.float64)
+                tf.TensorSpec(shape=(None), dtype=tf.float32)
             ]
         )(self._train_step)
     
@@ -354,7 +354,7 @@ class TextMelTransformer(Transformer):
                 tf.TensorSpec(shape=(None, None), dtype=tf.int64),
                 tf.TensorSpec(shape=(None, None, decoder_postnet.mel_channels), dtype=tf.float64),
                 tf.TensorSpec(shape=(None, None), dtype=tf.int64),
-                tf.TensorSpec(shape=(None,), dtype=tf.float64)
+                tf.TensorSpec(shape=(None), dtype=tf.float32)
             ]
         )(self._train_step)
